Add "try_get_for" for CompRepo
This returns an Option<&mut T>, which is 'None' if the given entity id
does not have a component in the repo (instead of panicking like
"get_for"). This avoids having to serach the repo once to check if the
entity has a component, and a second time to access the component.